chore: focus on Content-Range
content-type is not hard requirement, and implementations may have
easier or harder time with coming up with the right one.

if the content type info is not part of DAG metadata, and needs to me
sniffed (majority of content in 2025), then this internal sniffing may
produce different results in different file x library permutations.

in practice, it does not matter, browsers will stream videos just fine.

lets focus on Content-Range and body being correct - if we need to test
Content-Type, let's move it to a different test suite