test: range requests of deserialized files (#213)
* test: suffix range request

* chore: adjust range to not start at 0

makes test more representative of seeking inside of a partially
available file

* chore: focus on Content-Range

content-type is not hard requirement, and implementations may have
easier or harder time with coming up with the right one.

if the content type info is not part of DAG metadata, and needs to me
sniffed (majority of content in 2025), then this internal sniffing may
produce different results in different file x library permutations.

in practice, it does not matter, browsers will stream videos just fine.

lets focus on Content-Range and body being correct - if we need to test
Content-Type, let's move it to a different test suite

* chore: human-readable check results

* test: Accept-Ranges: bytes present

ensure behavior described in
https://github.com/ipfs/boxo/issues/856#issuecomment-2679171526
is covered by test

* chore: release as 0.8.0